move fuzzy search funcs; add basic test
