+++ fava/static/javascript/helpers.ts
@@ -1,4 +1,4 @@
-import { object, record, string, unknown } from "./validation";
+import { object, record, string, unknown } from "./lib/validation";
 import { favaAPI } from "./stores";
 
 /**
@@ -186,45 +186,4 @@ export async function putAPI<T extends keyof apiTypes>(
   }).then(handleJSON);
   const { data }: { data: unknown } = validateAPIResponse(res);
   return putAPIValidators[endpoint](data) as ReturnType<apiTypes[T]>;
-}
-
-/**
- * Fuzzy match a pattern against a string.
- *
- * Returns true if all characters of `pattern` can be found in order in
- * `string`. For lowercase characters in `pattern` match both lower and upper
- * case, for uppercase only an exact match counts.
- */
-export function fuzzytest(pattern: string, text: string): boolean {
-  let pindex = 0;
-  for (let index = 0; index < text.length; index += 1) {
-    const char = text[index];
-    const search = pattern[pindex];
-    if (char === search || char.toLowerCase() === search) {
-      pindex += 1;
-    }
-  }
-  return pindex === pattern.length;
-}
-
-/**
- * Wrap fuzzy matched characters.
- *
- * Wrap all occurences of characters of `pattern` (in order) in `string` in
- * <span> tags.
- */
-export function fuzzywrap(pattern: string, text: string): string {
-  let pindex = 0;
-  const result = [];
-  for (let index = 0; index < text.length; index += 1) {
-    const char = text[index];
-    const search = pattern[pindex];
-    if (char === search || char.toLowerCase() === search) {
-      result.push(`<span>${char}</span>`);
-      pindex += 1;
-    } else {
-      result.push(char);
-    }
-  }
-  return result.join("");
 }
